+++ django/applications/catmaid/static/js/init.js
@@ -809,6 +809,11 @@ function handle_message( status, text, xml )
 					}
 				}
 				message_menu.update( e );
+				// Make all message links open in a new page
+				var links = message_menu.getView().querySelectorAll('a');
+				for (var j=0; j<links.length; ++j) {
+					links[j].target = '_blank';
+				};
 				if ( n > 0 ) document.getElementById( "message_menu_text" ).className = "alert";
 				else document.getElementById( "message_menu_text" ).className = "";
 			}